size_t로 변경

+++ CarSimulator/Shader.cpp
@@ -633,7 +633,7 @@ void CLightsShader::UpdateShaderVariables(ID3D12GraphicsCommandList* pd3dCommand
 		m_pcbMappedLights[i] = VS_VB_LIGHT_INFO(m_vLight[i]);
 	}
 	
-	int vSize = m_vLight.size();
+	size_t vSize = m_vLight.size();
 	pd3dCommandList->SetGraphicsRootShaderResourceView(5, m_pd3dcbLight->GetGPUVirtualAddress());
 	pd3dCommandList->SetGraphicsRoot32BitConstants(2, 1, &vSize, 0);
 }
